+++ themes/blog/partials/copyright.php
@@ -1,0 +1,9 @@
+<div class="row g-0">
+    <div class="col text-center info-title">
+    <?php if (!is_null(SITE_COPYRIGHT) && "" != SITE_COPYRIGHT): ?>
+        <?php echo SITE_COPYRIGHT; ?>
+    <?php else: ?>
+        Powered by <a href="https://github.com/cwchentw/lightweight-cms" target="_blank" rel="noopener nofollow">Lightweight CMS</a>
+    <?php endif; ?>
+    </div>
+</div>